backend/requirements-staged.txt
# Install these packages one by one if you encounter issues
# Use: pip install <package>

# Core FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-dotenv==1.0.0
python-multipart==0.0.6

# MongoDB - Install these first
pymongo==4.6.0
motor==3.3.2

# Authentication - Install these second
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
email-validator==2.1.0

# LangChain - Install these third
langchain==0.3.26
langchain-community==0.3.26
langchain-google-genai==2.1.12

# Vector Database - Install after langchain
pinecone-client==3.2.2  # Use pinecone-client instead of pinecone
langchain-pinecone==0.2.0

# Embeddings
sentence-transformers==4.1.0

# Google Gemini
google-generativeai==0.8.5

# Document Processing - Install these last
pypdf==5.6.1
python-docx==1.1.0
openpyxl==3.1.2
aiofiles==23.2.1

# Utilities
httpx==0.28.1
